backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pandas==2.2.0
numpy==1.26.0
scikit-learn==1.4.0
scipy==1.12.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
duckdb==1.1.0
httpx==0.27.0
websockets>=12.0
